@pitdicker 2024-02-13
I rebased the documentation commit to be earlier in the PR, that should make reviewing a bit easier.

And there was one commit I had left after the `internals` refactor that makes the constants used in `YearFlags` less magical and updates the documentation. It kind of seems to fit in with this PR.